Finished the "generation of late notices" functionality, except for the automatic scheduling of the notice generation (Rich will do this), and the building of the actual pdfs. It currently produces a text file instead of a pdf.  This text file contains all the data required of the notice.
Updated the pages of the interface that allow editing of the 
Account table, so that it now handles the DisableLateNotices
property, properly.

Fixed a bug within the NewRateGroup and NewRate validation 
functionality.  This was with respect to using = to compare
the names of Rates and RateGroups, instead of using the LIKE
construct.

The list of adjustments shown on the invoices and payments page
now lists them in order of their ChargedOn date, not their
CreatedOn date