PDF Generation:
This is still a WIP and will be revisited to add helpful comments and remove debug comments. Some of the files are place-holders for work yet to be done. Some of the files will be removed altogether once development is complete.

The code added has a dependency on pear module Image_Barcode and the GD Image Library.

The Zend library code HAS BEEN MODIFIED to allow proper extension by subclasses. DO NOT REPLACE WITH THE VERSION SUPPLIED BY ZEND!

To allow for development testing the file /flex/pdf_dev.php has been added. When executed via the browser this should generate a PDF document in the /flex directory.