Modified roll-out script - Added dependency on PEAR::MDB2
Added support for new-style roll-outs which use the MDB2 PEAR library, work on Postgres databases and are applied incrementally.

Old-style roll-outs will still be applied in the old way, but the MDB2 library must still be installed; the new-style roll-out runs first and falls back to the old-style if non-applied roll-out scripts prior to 60 are found.